+++ packages/magic_ext_tezos/test/magic_ext_tezos_test.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:magic_ext_tezos/magic_ext_tezos.dart';
+
+void main() {}