+++ packages/magic_sdk/lib/magic_sdk.dart
@@ -1,4 +1,5 @@
 import 'modules/auth/auth_module.dart';
+import 'modules/blockchain/supported_blockchain.dart';
 import 'modules/user/user_module.dart';
 import 'modules/web3/eth_network.dart';
 import 'provider/rpc_provider.dart';
@@ -14,7 +15,7 @@ class Magic {
   /// Singleton proerty that can be accessed anywhere
   static late Magic instance;
 
-  /// rpcProvider that bounds to the instance
+  /// rpcProvider that routes payload for sign and send to the node
   late RpcProvider provider;
 
   /// Expose this property to display Webview
@@ -41,9 +42,20 @@ class Magic {
     URLBuilder.instance = URLBuilder.custom(apiKey, rpcUrl, chainId, locale);
     provider = RpcProvider(relayer);
   }
+
+  /// for custom node configuration provide chain id[String] and rpc url[String
+  Magic.blockchain(String apiKey,
+      {required String rpcUrl,
+      required SupportedBlockchain chain,
+      MagicLocale locale = MagicLocale.en_US}) {
+    URLBuilder.instance = URLBuilder.blockchain(apiKey, chain, rpcUrl, locale);
+    provider = RpcProvider(relayer);
+  }
 }
 
 /// Append basic modules on runtime
+/// so that core sdk does not depend on these these modules
+/// This is important to keep magic_sdk light weight
 extension MagicBaseModule on Magic {
   AuthModule get auth => AuthModule(provider);
   UserModule get user => UserModule(provider);