+++ packages/magic_sdk/lib/modules/auth/auth_method.dart
@@ -6,9 +6,3 @@ enum AuthMethod {
   magic_auth_login_with_sms,
   magic_auth_login_with_email_otp
 }
-
-extension ParseAuthMethodToString on AuthMethod {
-  String toShortString() {
-    return toString().split('.').last;
-  }
-}