+++ packages/magic_sdk/lib/modules/user/user_module.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 import 'dart:convert';
 
-import '../../magic_sdk.dart';
 import '../../modules/base_module.dart';
 import '../../modules/user/user_response_type.dart';
 import '../../provider/rpc_provider.dart';
@@ -20,7 +19,7 @@ class UserModule extends BaseModule {
       'lifespan': lifespan,
     };
     return sendToProvider(
-        method: UserMethod.magic_auth_get_id_token.toShortString(),
+        method: UserMethod.magic_auth_get_id_token,
         params: [params]).then((jsMsg) {
       var relayerResponse = RelayerResponse<String>.fromJson(
           json.decode(jsMsg.message), (json) => json as String);
@@ -33,7 +32,7 @@ class UserModule extends BaseModule {
       {int lifespan = 900, String attachment = 'none'}) async {
     var params = {'lifespan': lifespan, 'attachment': attachment};
     return sendToProvider(
-        method: UserMethod.magic_auth_generate_id_token.toShortString(),
+        method: UserMethod.magic_auth_generate_id_token,
         params: [params]).then((jsMsg) {
       var relayerResponse = RelayerResponse<String>.fromJson(
           json.decode(jsMsg.message), (json) => json as String);
@@ -43,8 +42,7 @@ class UserModule extends BaseModule {
 
   /// Returns [Future] of [UserMetadata], Retrieves information for the authenticated user.
   Future<UserMetadata> getMetadata() async {
-    return sendToProvider(
-            method: UserMethod.magic_auth_get_metadata.toShortString())
+    return sendToProvider(method: UserMethod.magic_auth_get_metadata)
         .then((jsMsg) {
       var relayerResponse = RelayerResponse<UserMetadata>.fromJson(
           json.decode(jsMsg.message),
@@ -55,8 +53,7 @@ class UserModule extends BaseModule {
 
   /// Returns [Future] of [bool], which denotes if user has already logged in, or not.
   Future<bool> isLoggedIn() async {
-    return sendToProvider(
-            method: UserMethod.magic_auth_is_logged_in.toShortString())
+    return sendToProvider(method: UserMethod.magic_auth_is_logged_in)
         .then((jsMsg) {
       var relayerResponse = RelayerResponse<bool>.fromJson(
           json.decode(jsMsg.message), (json) => json as bool);
@@ -70,7 +67,7 @@ class UserModule extends BaseModule {
     var params = {'email': email, 'showUI': showUI};
 
     return sendToProvider(
-        method: UserMethod.magic_auth_update_email.toShortString(),
+        method: UserMethod.magic_auth_update_email,
         params: [params]).then((jsMsg) {
       var relayerResponse = RelayerResponse<bool>.fromJson(
           json.decode(jsMsg.message), (json) => json as bool);
@@ -80,8 +77,7 @@ class UserModule extends BaseModule {
 
   /// Returns [Future] of [bool], Logs out the currently authenticated Magic user
   Future<bool> logout() async {
-    return sendToProvider(method: UserMethod.magic_auth_logout.toShortString())
-        .then((jsMsg) {
+    return sendToProvider(method: UserMethod.magic_auth_logout).then((jsMsg) {
       var relayerResponse = RelayerResponse<bool>.fromJson(
           json.decode(jsMsg.message), (json) => json as bool);
       return relayerResponse.response.result;