+++ packages/magic_sdk/lib/relayer/url_builder.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import '../../modules/web3/eth_network.dart';
 import '../../relayer/locale.dart';
+import '../modules/blockchain/supported_blockchain.dart';
 
 class URLBuilder {
   final String _host = "https://box.magic.link";
@@ -12,6 +13,7 @@ class URLBuilder {
   MagicLocale locale;
 
   Map? _network;
+  Map? _ext;
 
   String get encodedParams {
     var urlObj = {};
@@ -19,7 +21,14 @@ class URLBuilder {
     urlObj['host'] = _host;
     urlObj['sdk'] = 'magic-sdk-flutter';
     urlObj['locale'] = locale.toString().split('.').last;
-    urlObj['ETH_NETWORK'] = _network;
+
+    if (_network != null) {
+      urlObj['ETH_NETWORK'] = _network;
+    }
+
+    if (_ext != null) {
+      urlObj['ext'] = _ext;
+    }
 
     // Encode params to base64
     var jsonStr = json.encode(urlObj);
@@ -39,5 +48,23 @@ class URLBuilder {
 
   URLBuilder.eth(this.apiKey, EthNetwork network, this.locale) {
     _network = {"network": network.toString().split('.').last};
+  }
+
+  URLBuilder.blockchain(
+      this.apiKey, SupportedBlockchain chain, String rpcUrl, this.locale) {
+    String key;
+
+    /// Compatible blockchain name mapping
+    switch (chain) {
+      case SupportedBlockchain.tezos:
+        key = 'taquito';
+        break;
+      default:
+        key = chain.toString().split('.').last;
+    }
+
+    _ext = {
+      key: {"rpcUrl": rpcUrl}
+    };
   }
 }