+++ packages/magic_sdk/lib/utils/string.dart
@@ -1,0 +1,4 @@
+/// Used when converting enum
+String toShortString(Enum method) {
+  return method.toString().split('.').last;
+}